lvm: better handling of unknown segment types during parsing vg
grub_lvm_detect stopped incorrectly when an unknown segment type was found
that also contained a nested { } block.

Such an example is the 'cache-pool+METADATA_FORMAT" segment, which looks like
this:
segment1 {
        start_extent = 0
        extent_count = 7680     # 30 Gigabytes

        type = "cache-pool+METADATA_FORMAT"
[...]
        policy = "smq"
        policy_settings {
                migration_threshold=16384
        }
}

In the case the parser stopped at the type = " position, then seeked to the
first } character believing that this is the end of the block.

This patch fixes this issue by adding a new parser method that counts
the opening characters as well, and skips any nested blocks.

Signed-off-by: Nagy Attila Gabor <mrbig@sneaker.hu>